chore: rename rule and plugin to `no-empty-string`, update references and documentation accordingly
