+++ eslint.config.mjs
@@ -58,7 +58,7 @@ export default tsEslint.config([
       "sonarjs/no-element-overwrite": "error",
       "sonarjs/no-identical-conditions": "error",
       "sonarjs/no-identical-expressions": "error",
-      "@ubiquity-os/empty-string-checker": "warn",
+      "@ubiquity-os/no-empty-string": "warn",
       "@typescript-eslint/naming-convention": [
         "error",
         {
